[zephyr] Separate shared-pool execution responsibilities
Keep ZephyrContext as the only public pool lifecycle type. Move coordinator and worker actor code to separate modules. Store all owned actor groups in one private pool value.

Use one 32-thread result executor for each coordinator. Concurrent pipelines no longer make one executor for each pipeline.
